# src/api/contracts/school-openapi.yaml
swagger: "2.0"
info:
  title: School API
  description: School API
  version: 1.0.5
  contact:
    email: orscarsilla@outlook.com
...
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /sections:
    post:
      tags:
        - "Sections"
      summary: "Create section"
      description: "Create section"
      operationId: "createSection"
      parameters:
        - $ref: "#/parameters/sectionBodyParam"
      responses:
        '201':
          description: Created successfully.
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - "Sections"
      summary: "Get sections"
      description: "Get all sections"
      operationId: "getSections"
      responses:
        '200':
          description: OK.
          schema:
            $ref: "#/definitions/SectionsResponse"
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /sections/{ref}:
    get:
      tags:
        - "Sections"
      summary: "Get an existing section"
      description: "Get an existing section by ObjectId."
      operationId: "getSection"
      parameters:
        - $ref: "#/parameters/sectionReference"
      responses:
        '200':
          description: OK.
          schema:
            $ref: '#/definitions/SectionResponse'
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      tags:
        - "Sections"
      summary: "Modify an existing section"
      description: "Modify an existing section by ObjectId"
      operationId: "modifySection"
      parameters:
        - $ref: "#/parameters/sectionReference"
        - $ref: "#/parameters/sectionBodyParam"
      responses:
        '200':
          description: OK.
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - "Sections"
      summary: "Delete an existing section"
      description: "Delete an existing section by ObjectId"
      operationId: "deleteSection"
      parameters:
        - $ref: "#/parameters/sectionReference"
      responses:
        '204':
          description: No content.
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /articles:
    get:
      tags:
        - "Articles"
      summary: "Get articles"
      description: "Get all articles"
      operationId: "getArticles"
      parameters:
        - $ref: "#/parameters/limitParam"
        - $ref: "#/parameters/pageParam"
      responses:
        '200':
          description: OK.
          schema:
            $ref: "#/definitions/ArticlesResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - "Articles"
      summary: "Create article"
      description: "Create article"
      operationId: "createArticle"
      parameters:
        - $ref: "#/parameters/newBodyParam"
      responses:
        '201':
          description: Created.
          schema:
            $ref: "#/definitions/GeneratedIdResponse"
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /articles/{id}:
    get:
      tags:
        - "Articles"
      summary: "Get article"
      description: "Get article by id"
      operationId: "getArticle"
      parameters:
        - $ref: "#/parameters/idParam"
      responses:
        '200':
          description: OK.
          schema:
            $ref: "#/definitions/ArticleResponse"
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      tags:
        - "Articles"
      summary: "Update article"
      description: "Update article by id"
      operationId: "updateArticle"
      parameters:
        - $ref: "#/parameters/idParam"
        - $ref: "#/parameters/newBodyParam"
      responses:
        '200':
          description: OK.
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - "Articles"
      summary: "Delete article"
      description: "Delete article by id"
      operationId: "deleteArticle"
      parameters:
        - $ref: "#/parameters/idParam"
      responses:
        '204':
          description: No content.
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /events:
    get:
      tags:
        - "Events"
      summary: "Get events"
      description: "Get all events with optional pagination"
      operationId: "getEvents"
      parameters:
        - $ref: "#/parameters/limitParam"
        - $ref: "#/parameters/pageParam"
      responses:
        '200':
          description: OK.
          schema:
            $ref: "#/definitions/EventsResponse"
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - "Events"
      summary: "Create event"
      description: "Create new event"
      operationId: "createEvent"
      parameters:
        - $ref: "#/parameters/eventBodyParam"
      responses:
        '201':
          description: Created successfully.
          schema:
            $ref: "#/definitions/GeneratedIdResponse"
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /events/{id}:
    get:
      tags:
        - "Events"
      summary: "Get event"
      description: "Get an existing event by _id"
      operationId: "getEvent"
      parameters:
        - $ref: "#/parameters/idParam"
      responses:
        '200':
          description: OK.
          schema:
            $ref: "#/definitions/EventResponse"
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      tags:
        - "Events"
      summary: "Modify event"
      description: "Modify an existing event by _id"
      operationId: "modifyEvent"
      parameters:
        - $ref: "#/parameters/idParam"
        - $ref: "#/parameters/eventBodyParam"
      responses:
        '200':
          description: OK.
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - "Events"
      summary: "Delete event"
      description: "Delete an existing event by _id"
      operationId: "deteleEvent"
      parameters:
        - $ref: "#/parameters/idParam"
      responses:
        '204':
          description: No content.
        '400':
          description: Bad request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: Not found.
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  EventBody:
    type: object
    required:
      - "title"
      - "img"
      - "description"
    properties:
      title:
        type: string
        format: string
        example: "Event title"
      img:
        type: string
        format: string
        example: "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/"
      description:
        type: string
        format: string
        example: "Event description"
      content:
        type: string
        format: string
        example: <h1>Event content</h1>
  EventsResponse:
    type: array
    items:
      $ref: "#/definitions/EventResponse"
  EventResponse:
    type: object
    properties:
      id:
        type: string
        format: string
        example: "507f191e810c19729de860ea"
      title:
        type: string
        format: string
        example: "Event title"
      description:
        type: string
        format: string
        example: "Event description"
      img:
        type: string
        format: string
        example: "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/"
      content:
        type: string
        format: string
        example: "<h1>Event content</h1>"
  ArticleBody:
    type: object
    required:
      - "title"
      - "content"
    properties:
      title:
        type: string
        format: string
        example: "First new"
      description:
        type: string
        format: string
        example: "Article description"
      img:
        type: string
        format: string
        example: "portrait.jpg"
      content:
        type: string
        format: string
        example: "<h1>Content</h1>"
  ArticlesResponse:
    type: array
    items:
      $ref: "#/definitions/ArticleResponse"
  ArticleResponse:
    type: object
    required:
      - "_id"
      - "title"
      - "content"
    properties:
      _id:
        type: string
        format: string
        example: "507f191e810c19729de860ea"
      title:
        type: string
        format: string
        example: "First new"
      description:
        type: string
        format: string
        example: "Article description"
      img:
        type: string
        format: string
        example: "portrait.jpg"
      content:
        type: string
        format: string
        example: "<h1>Content</h1>"
...
  ErrorResponse:
    type: object
    required:
      - "code"
      - "description"
    properties:
      code:
        type: integer
        format: int32
      description:
        type: string
        format: string
...
  SectionBody:
    type: object
    required:
      - "title"
    properties:
      title:
        type: string
        format: string
        example: "Educational Principles"
      description:
        type: string
        format: string
        example: "Educational principles are the core of the school."
      img:
        type: string
        format: string
        example: "example.png"
...
  SectionsResponse:
    type: array
    items:
      $ref: "#/definitions/SectionResponse"
  SectionResponse:
    type: object
    properties:
      _id:
        type: string
        format: string
        example: "507f191e810c19729de860ea"
      title:
        type: string
        format: string
        example: "Educational Principles"
      description:
        type: string
        format: string
        example: "Educational principles are the core of the school."
      img:
        type: string
        format: string
        example: "example.png"
      ref:
        type: string
        format: string
        example: "educational_principles"
...
  GeneratedIdResponse:
    type: object
    properties:
      generatedId:
        type: string
        format: string
        example: "63b04cf26cff5203de1659ca"
parameters:
  idParam:
    name: "_id"
    description: "ObjectId param."
    in: path
    required: true
    type: string
    format: string
  sectionReference:
    name: "ref"
    description: "Section reference filter to find, update and delete a section."
    in: path
    required: true
    type: string
    format: string
...
  sectionBodyParam:
    name: "body"
    description: "Section body"
    in: body
    required: true
    schema:
      $ref: "#/definitions/SectionBody"
...
  newBodyParam:
    name: "body"
    description: "Article body param"
    in: body
    required: true
    schema:
      $ref: "#/definitions/ArticleBody"
  eventBodyParam:
    name: "body"
    description: "Event body param"
    in: body
    required: true
    schema:
      $ref: "#/definitions/EventBody"
  limitParam:
    name: "limit"
    description: "Limit of records for page"
    in: query
    type: string
    format: string
  pageParam:
    name: "page"
    description: "Start page for start pagination"
    in: query
    type: string
    format: string